+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2024-10-06
+
+### Added
+
+- `CsvColumnIndexAttribute` to be able to map you CSV by column index.
+
+- Tests for `CsvColumnIndexAttribute`.
+
+### Changed
+
+- Change parameter `header` name to `hasHeader`.
+
+- Make `hasHeader` method parameter non-optional.
+
 ## [1.1.1] - 2024-10-06
 
 ### Breaking changes
@@ -12,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Renamed the primary static class to `CsvParser`.
 
-## Changed
+### Changed
 
 - Moved some parts of the code to their own utility classes and files.